+++ Realtime/src/commonMain/kotlin/io/github/jan/supabase/realtime/RealtimeChannelBuilder.kt
@@ -1,7 +1,7 @@
 package io.github.jan.supabase.realtime
 
 import io.github.jan.supabase.annotations.SupabaseInternal
-import io.github.jan.supabase.realtime.annotiations.ChannelDsl
+import io.github.jan.supabase.realtime.annotations.ChannelDsl
 
 /**
  * Used to build a realtime channel
